consumer/requirements.txt
kafka-python==2.0.2
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
